Update assignment_1.py
Question-1:-Help to Mr X to bake the code. Mr X wants to find the occurrences of vowels in a
sentence and how many times each vowel repeated in this sentence. Mr X will have power to define
this sentence.
Input:-
Input a sentence from user.
Output:-
Vowels in sentence---
Each vowel repeated as:-
Example-1:-
Input:-
Mr X is working in Grads IT.
Output:- vowels in sentence—[A,E, I,O]
Each vowel repeated as—[1, 0, 4,1] (request with vowel)
Description:-- In the sentence “Mr X is working in Grads IT.” vowel A,E,I,O are present.
And Vowel A is repeated 1 time, Vowel E is repeated 0 time and vowel I repeated 4 times
and vowel O is repeated 1 time.