Add files via upload
TEA/ Coffee Vending Machine
The Mr. X has a restaurant and he wants to install an automatic vending
Machine for TEA /Coffee. The TEA / Coffee ingredients can be selected
according to the customer's choice. Machine also allows to store the
information of ingredients combination tagged with customer names and
prepare TEA / Coffee by entering their names.
Mr. X contacts a developer like you and wants a software that fully fulfills the
above requirements so the machine works accordingly. You can assume the
required functionality which helps to develop the software for Mr. X.
You are advised to design this software with a fully object oriented paradigm
(OOPs). The identified Objects and their required fields and methods totally
depend on the developer's choice.